Tests/reorganize: move email transport tests to own file
